Move Pupil Manager into separate file
